+++ heat/core/dndarray.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import numpy as np
 import math
 import torch
@@ -359,12 +357,16 @@ class DNDarray:
             # only exchange data with next process if it has data
             if rank != last_rank and (lshape_map[next_rank, self.split] > 0):
                 self.comm.Isend(a_next, next_rank)
-                res_prev = torch.zeros(a_prev.size(), dtype=a_prev.dtype)
+                res_prev = torch.zeros(
+                    a_prev.size(), dtype=a_prev.dtype, device=self.device.torch_device
+                )
                 req_list.append(self.comm.Irecv(res_prev, source=next_rank))
 
             if rank != 0:
                 self.comm.Isend(a_prev, prev_rank)
-                res_next = torch.zeros(a_next.size(), dtype=a_next.dtype)
+                res_next = torch.zeros(
+                    a_next.size(), dtype=a_next.dtype, device=self.device.torch_device
+                )
                 req_list.append(self.comm.Irecv(res_next, source=prev_rank))
 
             for req in req_list:
@@ -1425,6 +1427,8 @@ class DNDarray:
             """ if the key is a DNDarray and it has as many dimensions as self, then each of the entries in the 0th
                 dim refer to a single element. To handle this, the key is split into the torch tensors for each dimension.
                 This signals that advanced indexing is to be used. """
+            key.balance_()
+            key = manipulations.resplit(key.copy())
             lkey = [slice(None, None, None)] * self.ndim
             kgshape_flag = True
             kgshape = [0] * len(self.gshape)
@@ -1442,6 +1446,8 @@ class DNDarray:
                 lists mean advanced indexing will be used"""
             h = [slice(None, None, None)] * self.ndim
             if isinstance(key, DNDarray):
+                key.balance_()
+                key = manipulations.resplit(key.copy())
                 h[0] = key._DNDarray__array.tolist()
             elif isinstance(key, torch.Tensor):
                 h[0] = key.tolist()